requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
lxml==6.0.2
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
